+++ angular/src/app/data/address.model.ts
@@ -1,0 +1,20 @@
+/**
+ * Represents the _Address_ model
+ *
+ * ```yaml
+ * id: string;
+ * city: string;
+ * country: string;
+ * postalCode: string;
+ * stateProvince: string;
+ * street: string;
+ * ```
+ */
+export interface Address {
+  id: string;
+  city: string;
+  country: string;
+  postalCode: string;
+  stateProvince: string;
+  street: string;
+}