+++ angular/src/app/data/config.model.ts
@@ -1,0 +1,14 @@
+/**
+ * Represents the _Config_ model
+ *
+ * ```yaml
+ * api: object;
+ * ```
+ */
+export interface Config {
+  api: {
+    account: string;
+    booking: string;
+    lodging: string;
+  };
+}