+++ angular/src/app/data/payment.model.ts
@@ -1,0 +1,16 @@
+/**
+ * Represents the _Payment_ model
+ *
+ * ```yaml
+ * id: string;
+ * cardExpirationDate: string;
+ * cardName: string;
+ * cardNumber: string;
+ * ```
+ */
+export interface Payment {
+  id: string;
+  cardExpirationDate: string;
+  cardName: string;
+  cardNumber: string;
+}