+++ angular/src/app/services/account/account.service.ts
@@ -1,9 +1,62 @@
+import { HttpClient, HttpParams } from '@angular/common/http';
 import { Injectable } from '@angular/core';
+import { Observable } from 'rxjs';
+import { concatMap, map } from 'rxjs/operators';
+import { ConfigService } from '../config/config.service';
+import { Account } from '../../data/account.model';
 
 @Injectable({
-  providedIn: 'root'
+  providedIn: 'root',
 })
 export class AccountService {
+  private readonly apiUrl$: Observable<string>;
 
-  constructor() { }
+  /**
+   * Represents the _Account Service_ `constructor` method
+   *
+   * @param config ConfigService
+   * @param http HttpClient
+   */
+  constructor(private readonly config: ConfigService, private readonly http: HttpClient) {
+    this.apiUrl$ = config.get().pipe(map((cfg) => cfg.api.account));
+  }
+
+  /**
+   * Represents the _Account Service_ `delete` method
+   *
+   * @param id string
+   */
+  delete(id: string): Observable<boolean> {
+    return this.apiUrl$.pipe(
+      concatMap((url) => this.http.delete<boolean>(url, { params: { id } }))
+    );
+  }
+
+  /**
+   * Represents the _Account Service_ `get` method
+   *
+   * @param id string
+   */
+  get(id?: string): Observable<Account[]> {
+    const options = id ? { params: new HttpParams().set('id', id) } : {};
+    return this.apiUrl$.pipe(concatMap((url) => this.http.get<Account[]>(url, options)));
+  }
+
+  /**
+   * Represents the _Account Service_ `post` method
+   *
+   * @param account Account
+   */
+  post(account: Account): Observable<boolean> {
+    return this.apiUrl$.pipe(concatMap((url) => this.http.post<boolean>(url, account)));
+  }
+
+  /**
+   * Represents the _Account Service_ `put` method
+   *
+   * @param account Account
+   */
+  put(account: Account): Observable<Account> {
+    return this.apiUrl$.pipe(concatMap((url) => this.http.put<Account>(url, account)));
+  }
 }